feat(tags): fix tag page, update tag display to use tag_handle instead of name
Overview:
Fix the tag selection by replacing the use of `name` with `tag_handle` for better clarity and consistency in the user interface.

Changes:
- Modifies the community description section to show `tag_handle` for the selected tag.

This was updated because we standarized both `user` and `tag` documents to both use "handle" as a standarized description for this field, using the same validation functions for both. Now it is `tag_handle` and `user_handle`.